generated_simple_ctle_msdsl/requirements.txt
scikit-rf==0.17.0 # needed because the latest scikit-rf version is buggy
msdsl==0.3.7
PyYAML==5.4.1